% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-mase-scale-feat.R
\name{add_mase_scale_feat}
\alias{add_mase_scale_feat}
\title{Add a scale feature by using a factor derived from the MASE error function}
\usage{
add_mase_scale_feat(.tbl, .value, ...)
}
\arguments{
\item{.tbl}{tibble; data with grouping column and value column}

\item{.value}{numeric; unquoted column name that contains the numeric values of the time series}

\item{...}{character; one or more unquoted grouping columns}
}
\value{
The original tibble with an additional column, "scale."
}
\description{
\code{\link[=add_mase_scale_feat]{add_mase_scale_feat()}} calculates a MASE scale factor and divides this factor by the group average scale factor to produce a scale feature.
}
\details{
Designed to use with a global forecasting method. It's recommended to standardize the stacked series that is used as input for this method. Standardizing the stacked series removes the scale information about each series in the stack which might be useful in generating the forecast. Adding a scale feature reintroduces this information back into the model.
}
\examples{

library(dplyr, warn.conflicts = FALSE)

group_ts_tbl <- tsbox::ts_tbl(fpp2::arrivals)

head(group_ts_tbl)

new_tbl <- add_mase_scale_feat(group_ts_tbl, .value = value, id)

head(new_tbl)
}
\references{
Pablo Montero-Manso, Rob J. Hyndman, Principles and algorithms for forecasting groups of time series: Locality and globality, International Journal of Forecasting, 2021 \href{https://robjhyndman.com/publications/global-forecasting/}{link}
}
