% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-tsfeatures.R
\name{add_tsfeatures}
\alias{add_tsfeatures}
\title{Add time series features calculated by the tsfeatures package}
\usage{
add_tsfeatures(.tbl, ..., standardize = TRUE, parallel = FALSE)
}
\arguments{
\item{.tbl}{tibble; data with date (class: Date), value (class: numeric), and group (class: character) columns}

\item{...}{character; one or more unquoted grouping columns}

\item{standardize}{logical; If TRUE (default), the function with standardize each feature.}

\item{parallel}{logical; If TRUE, features will be calculated in parallel. Default is FALSE.}
}
\value{
The original tibble with 20 additional feature columns.
}
\description{
\code{\link[=add_tsfeatures]{add_tsfeatures()}} adds a set of calculated features from the \href{https://pkg.robjhyndman.com/tsfeatures/}{tsfeatures} package for each time series in the group. These features provide information about various characteristics of the time series.
}
\details{
Function can be used with a global forecasting method or for EDA. See the \href{https://pkg.robjhyndman.com/tsfeatures/}{tsfeatures} website for more details on these features.
}
\examples{

library(dplyr, warn.conflicts = FALSE)

group_ts_tbl <- tsbox::ts_tbl(fpp2::arrivals)

head(group_ts_tbl)

new_tbl <- add_tsfeatures(group_ts_tbl, id)

head(new_tbl)
}
\references{
Pablo Montero-Manso, Rob J. Hyndman, Principles and algorithms for forecasting groups of time series: Locality and globality, International Journal of Forecasting, 2021 \href{https://robjhyndman.com/publications/global-forecasting/}{link}
}
\seealso{
\code{\link[tsfeatures:tsfeatures]{tsfeatures::tsfeatures()}}
}
