% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cles.R
\name{cles}
\alias{cles}
\title{Calculates the Common Language Effect Size (CLES)}
\usage{
cles(data, variable, group, baseline)
}
\arguments{
\item{data}{dataframe; Data should be in long format}

\item{variable}{numeric vector; Variable of interest}

\item{group}{string; The name of group variable}

\item{baseline}{string; The name of the category of the group variable that's the control group or baseline in a repeated measure scenario.}
}
\value{
The function returns a scalar value.
}
\description{
Calculates the Common Language Effect Size (CLES) for two variables. The CLES function converts the effect size to a probability that a unit/subject will have a larger measurement than another unit/subject. See \href{https://ercbk.github.io/Data-Science-Notebook/qmd/post-hoc-analysis-multilevel.html#common-language-effect-size}{notebook} for further details.
}
\details{
This measure is also referred to as the \emph{Probability of Superiority}. The conversion of effect size to a probability or percentage is supposed to be easier for the laymen to interpret.
Interpretation:
\itemize{
\item Between-Subjects: The probability that a randomly sampled person from one group will have a higher observed measurement than a randomly sampled person from the other group.
\item Within-Subjects: The probability that an individual has a higher value on one measurement than the other.
}

Formula:
\deqn{\tilde d = \frac{|M_1 - M_2|}{\sqrt{p_1\text{SD}_1^2 + p_2\text{SD}_2^2}}\\ Z = \frac{\tilde d}{\sqrt{2}}}
\itemize{
\item \eqn{M_i}: The mean of the i\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} group
\item \eqn{p_i}: The proportion of the sample size of the i\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} group
\item \eqn{Z}: The z-score which is in turn used to produce the probability.
\item Same formula is used for both within and between designs
}
}
\examples{

movie_dat <- dplyr::tibble(
   movie1 = c(9.00, 7.00, 8.00, 9.00, 8.00, 9.00, 9.00, 10.00, 9.00, 9.00),
   movie2 = c(9.00, 6.00, 7.00, 8.00, 7.00, 9.00, 8.00, 8.00, 8.00, 7.00)
)

movie_dat |>
 tidyr::pivot_longer(cols = c(movie1, movie2),
                     names_to = "movies",
                     values_to ="ratings") |>
 cles("ratings", "movies", "movie1")
}
\references{
McGraw, K. O., & Wong, S. P. (1992). A common language effect size statistic. Psychological Bulletin, 111(2), 361–365. \url{https://doi.org/10.1037/0033-2909.111.2.361}
}
