% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-dtw-grids.R
\name{create_dtw_grids}
\alias{create_dtw_grids}
\title{Create a parameter grid list for \code{dtwclust} and \code{dtw} distance functions}
\usage{
create_dtw_grids(params)
}
\arguments{
\item{params}{A named list of parameter name-value pairs. The names should be the names of the distance functions that correspond to the parameter name-value pairs.}
}
\value{
Named, nested list where each element is a list of elements that represent each possible configuration for the parameters of that distance function.
}
\description{
\code{\link[=create_dtw_grids]{create_dtw_grids()}} creates a nested, parameter grid list for dtwclust and dtw distance functions to be used in \code{\link[=dtw_dist_gridsearch]{dtw_dist_gridsearch()}}.
}
\examples{

params_ls_lg <- list(dtw_basic = list(window.size = 5:10,
                                      norm = c("L1", "L2"),
                                      step.pattern = list(dtw::symmetric1, dtw::symmetric2)),
                     dtw2 = list(step.pattern = list(dtw::symmetric1, dtw::symmetric2),
                                 window.size = 5:10),
                     dtw_lb = list(window.size = 5:10,
                                   norm = c("L1", "L2"),
                                   dtw.func = "dtw_basic",
                                   step.pattern = list(dtw::symmetric2)),
                     sbd = list(znorm = TRUE, return.shifted = FALSE),
                     gak = list(normalize = TRUE, window.size = 5:10))

dtw_grids_lg <- create_dtw_grids(params_ls_lg)
head(dtw_grids_lg, n = 2)


# Can still be ran with a minimal "grid"
params_ls_sm <- list(dtw2 = list(step.pattern = list(dtw::symmetric1)))

dtw_grids_sm <- create_dtw_grids(params_ls_sm)
head(dtw_grids_sm)
}
