% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descale-by-mase.R
\name{descale_by_mase}
\alias{descale_by_mase}
\title{Back-transform a MASE-scaled column}
\usage{
descale_by_mase(.tbl, .value, scale_factors, ...)
}
\arguments{
\item{.tbl}{tibble; data with a value (class: numeric) column and group (class: character) column(s)}

\item{.value}{numeric; unquoted name of the column that contains the scaled numeric values}

\item{scale_factors}{tibble; tibble extracted from the \code{scale_factors} attribute of the output of \code{\link[=scale_by_mase]{scale_by_mase()}}. Grouping columns should match those in \code{.tbl}. (See details)}

\item{...}{character; one or more unquoted grouping columns}
}
\value{
The original tibble with the \code{.value} column back-transformed to the orginal scale.
}
\description{
\code{\link[=descale_by_mase]{descale_by_mase()}} back-transforms a group time series that has been scaled by a factor derived from the MASE error function.
}
\details{
Scaling a grouped time series can be helpful for global forecasting methods when using machine learning and deep learning algorithms. Scaling by MASE and using MASE as the error function is equivalent to to minimizing the MAE in the preprocessed time series.

The \code{scale_factors} tibble can be extracted by \code{scale_factors <- attributes(mase_scaled_tbl)$scale_factors} where \code{mase_scaled_tbl} is the output of \code{\link[=scale_by_mase]{scale_by_mase()}}.
}
\examples{

library(dplyr, warn.conflicts = FALSE)

group_ts_tbl <- tsbox::ts_tbl(fpp2::arrivals)

head(group_ts_tbl)

new_tbl <- scale_by_mase(.tbl = group_ts_tbl, .value = value, id)

glimpse(new_tbl)

scale_factors <- attributes(new_tbl)$scale_factors

orig_tbl <- descale_by_mase(new_tbl, value, scale_factors, id)

head(orig_tbl)
}
\references{
Pablo Montero-Manso, Rob J. Hyndman, Principles and algorithms for forecasting groups of time series: Locality and globality, International Journal of Forecasting, 2021 \href{https://robjhyndman.com/publications/global-forecasting/}{link}
}
