% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indiana_pos_rate}
\alias{indiana_pos_rate}
\title{Indiana COVID-19 Positivity Rates Data Set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 528 rows and 3 columns.
}
\usage{
indiana_pos_rate
}
\description{
Contains weekly COVID-19 positivity rates for metropolitan statistical areas in Indiana from April 4th, 2020 to January 30th, 2021.
}
\details{
Data collected from state health departments and curated in my \href{https://github.com/ercbk/Indiana-COVIDcast-Dashboard/tree/master/data}{Indiana COVIDcast Dashboard} repository.
}
\examples{
head(indiana_pos_rate)
}
\keyword{datasets}
