% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prewhitened-ccf.R
\name{prewhitened_ccf}
\alias{prewhitened_ccf}
\title{Calculate cross-correlation coefficients for prewhitened time series}
\usage{
prewhitened_ccf(
  input,
  output,
  input_col,
  output_col,
  keep_input = "both",
  keep_ccf = "both",
  max.order
)
}
\arguments{
\item{input}{tsibble; The influential or "predictor-like" time series}

\item{output}{tsibble; The affected or "response-like" time series}

\item{input_col}{string; Name of the numeric column of the input tsibble}

\item{output_col}{string; Name of the numeric column of the output tsibble}

\item{keep_input}{string; values: "input_lags", "input_leads" or "both"; Default is "both".}

\item{keep_ccf}{string; values: "positive", "negative", or "both; Default is "both"}

\item{max.order}{integer; The maximum lag used in the CCF calculation.}
}
\value{
A tibble with the following columns:
\itemize{
\item input_type: "lag" or "lead"
\item input_series: lag or lead number
\item signif_type: "Statistically Significant" or "Not Statistically Significant"
\item signif_threshold: Threashold CCF value for statistical significance at the 95\% level
\item ccf: Calculated ccf value
}
}
\description{
\code{\link[=prewhitened_ccf]{prewhitened_ccf()}} prewhitens time series, calculates cross-correlation coefficients, and returns statistically significant values.
}
\details{
In a cross-correlation in which the direction of influence between two time-series is hypothesized or known,
\itemize{
\item the influential time-series is called the "input" time-series
\item the affected time-series is called the "output" time-series
}

The cross-correlation function calculates correlation values between lags and leads of the input series and the output series. Sometimes only correlations between the leads or lags of the input series and the output series make theoretical sense, or only positive or negative correlations make theoretical sense.
\itemize{
\item The "keep_input" argument specifies whether you want to keep only output CCF values involving leads or lags of the input series or both.
\item The "keep_ccf" argument specifies whether you want to keep only output positive, negative, or both CCF values.
}

The prewhitening method that is used is from Cryer and Chan (2008, Chapter 11) Time Series Analysis. \code{prewhitened_ccf} differences the series if it's needed and outputs either statistically significant values of the CCF or the top non-statistically significant value if no statistically significant values are found.
}
\examples{

oh_cases <- ohio_covid \%>\%
   dplyr::select(date, cases) \%>\%
   tsibble::as_tsibble(index = date)

oh_deaths <- ohio_covid \%>\%
   dplyr::select(date, deaths_lead0) \%>\%
   tsibble::as_tsibble(index = date)

oh_ccf_tbl <- prewhitened_ccf(input = oh_cases,
                              output = oh_deaths,
                              input_col = "cases",
                              output_col = "deaths_lead0",
                              max.order = 40,
                              keep_input = "input_lag",
                              keep_ccf = "positive")

oh_ccf_tbl
}
\references{
Cryer, D., Chan, K. (2008) Time Series Analysis, Springer Science+Business Media, LLC
}
