% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-by-mase.R
\name{scale_by_mase}
\alias{scale_by_mase}
\title{Scale a group time series by using a factor derived from the MASE error function}
\usage{
scale_by_mase(.tbl, .value, ...)
}
\arguments{
\item{.tbl}{tibble; data with a value (class: numeric) column and group (class: character) column(s)}

\item{.value}{numeric; unquoted name of the column that contains the scaled numeric values}

\item{...}{character; one or more unquoted grouping columns}
}
\value{
The original tibble with the \code{.value} column back-transformed to the orginal scale.
}
\description{
\code{\link[=scale_by_mase]{scale_by_mase()}} scales a group time series by using a factor derived from the MASE error function.
}
\details{
Scaling a grouped time series can be helpful for global forecasting methods when using for machine learning and deep learning algorithms. Scaling by MASE and using MASE as the error function is equivalent to to minimizing the MAE in the preprocessed time series.

For each series, a MASE scale factor is calculated using the denominator of the MASE scaled error equation. Then, the series is divided by this factor.
}
\examples{

library(dplyr, warn.conflicts = FALSE)

group_ts_tbl <- tsbox::ts_tbl(fpp2::arrivals)

head(group_ts_tbl)

new_tbl <- scale_by_mase(.tbl = group_ts_tbl, .value = value, id)

head(new_tbl)

attributes(new_tbl)$scale_factors
}
\references{
Pablo Montero-Manso, Rob J. Hyndman, Principles and algorithms for forecasting groups of time series: Locality and globality, International Journal of Forecasting, 2021 \href{https://robjhyndman.com/publications/global-forecasting/}{link}
}
