% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_regional_deaths}
\alias{us_regional_deaths}
\title{US Regional COVID-19 Deaths Data Set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1276 rows and 3 columns.
}
\usage{
us_regional_deaths
}
\description{
Contains rolling 7-day average COVID-19 deaths for US regional areas from April 1st, 2020 to February 13th, 2021.
}
\details{
Data collected by the \href{https://github.com/cmu-delphi/covidcast}{Delphi Research Group}.
}
\examples{
head(us_regional_deaths)
}
\keyword{datasets}
