% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cles.R
\name{cles}
\alias{cles}
\title{Calculates the Common Language Effect Size (CLES)}
\usage{
cles(data, group_variables, paired = FALSE, ci = FALSE, ...)
}
\arguments{
\item{data}{dataframe; Data should be in wide format}

\item{group_variables}{character vector or list with quoted names of the variables to be compared.}

\item{paired}{boolean; Indicates whether variables are correlated as in a repeated measures design. Default is FALSE.}

\item{ci}{boolean; Indicates whether bootstrap confidence intervals should be calculated. Default is FALSE.}

\item{...}{Additional arguments that should be passed to \code{\link[=get_boot_ci]{get_boot_ci()}}}
}
\value{
When 'ci = FALSE', this function returns a scalar value estimate of the CLES.
When 'ci = TRUE', this function returns a dataframe with the following columns:
\itemize{
\item ci_type: The method of calculating the bootstrap confidence intervals.
\item conf: The confidence level for the bootstrap confidence intervals,
\item .lower: The lower value of the bootstrap confidence interval.
\item .estimate: The CLES point estimate.
\item .upper: The upper value of the bootstrap confidence interval.
}
}
\description{
Calculates the Common Language Effect Size (CLES) for two variables. The CLES function converts the effect size to a probability that a unit/subject will have a larger measurement than another unit/subject. See \href{https://ercbk.github.io/Data-Science-Notebook/qmd/post-hoc-analysis-multilevel.html#common-language-effect-size}{notebook} for further details.
}
\details{
This measure is also referred to as the \emph{Probability of Superiority}. The conversion of effect size to a probability or percentage is supposed to be easier for the laymen to interpret.
Interpretation:
\itemize{
\item Between-Subjects: The probability that a randomly sampled person from one group will have a higher observed measurement than a randomly sampled person from the other group.
\item Within-Subjects: The probability that an individual has a higher value on one measurement than the other.
}

Between-Subjects Formula:
\deqn{\tilde d = \frac{|M_1 - M_2|}{\sqrt{p_1\text{SD}_1^2 + p_2\text{SD}_2^2}}\\ Z = \frac{\tilde d}{\sqrt{2}}}
\itemize{
\item \eqn{M_i}: The mean of the i\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} group
\item \eqn{p_i}: The proportion of the sample size of the i\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} group
\item \eqn{Z}: The z-score which is in turn used to produce the probability.
}

Within-Subjects Formula:
\deqn{Z = \frac{|M_1 - M_2|}{\sqrt{\operatorname{SD}_1^2 + \operatorname{SD}_2^2 - 2 \times r \times \operatorname{SD}_1 \times \operatorname{SD}_2}}}
\itemize{
\item \eqn{M_i}: The mean of the i\if{html}{\out{<sup>}}th\if{html}{\out{</sup>}} group
\item \eqn{r}: Pearson correlation between the two variables
\item \eqn{Z}: The z-score which is in turn used to produce the probability.
}
}
\examples{

movie_dat <- dplyr::tibble(
   movie1 = c(9.00, 7.00, 8.00, 9.00, 8.00, 9.00, 9.00, 10.00, 9.00, 9.00),
   movie2 = c(9.00, 6.00, 7.00, 8.00, 7.00, 9.00, 8.00, 8.00, 8.00, 7.00)
)

# between-subjects design
cles(data = movie_dat,
     group_variables = list("movie1", "movie2"))

# within-subjects design and bootstrap CIs
cles(data = movie_dat,
     group_variables = list("movie1", "movie2"),
     paired = TRUE,
     ci = TRUE,
     R = 10000,
     type = c("bca", "perc"))
}
\references{
McGraw, K. O., & Wong, S. P. (1992). A common language effect size statistic. Psychological Bulletin, 111(2), 361–365. \url{https://doi.org/10.1037/0033-2909.111.2.361}
}
