% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ohio_covid}
\alias{ohio_covid}
\title{Ohio COVID-19 Data Set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 255 rows and 63 columns.
}
\usage{
ohio_covid
}
\description{
Contains rolling 7-day averages of positive COVID-19 cases and leads of COVID-19 deaths for Ohio from March 22nd, 2020 to December 1st, 2020.
}
\details{
Data collected from the New York Times COVID-19 data \href{https://github.com/nytimes/covid-19-data}{repository}
}
\examples{
head(ohio_covid)[,1:6]
}
\keyword{datasets}
